+++ common/src/main/java/io/github/mortuusars/envelope/client/gui/tooltip/Tooltip.java
@@ -1,7 +1,10 @@
 package io.github.mortuusars.envelope.client.gui.tooltip;
 
 import io.github.mortuusars.envelope.Envelope;
+import io.github.mortuusars.envelope.world.mail.address.Address;
 import io.github.mortuusars.envelope.world.mail.address.AddressFormatter;
+import net.minecraft.ChatFormatting;
+import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.Nullable;
@@ -15,12 +18,25 @@ public class Tooltip {
                 return;
             }
 
-            @Nullable Component senderToRecipient = AddressFormatter.senderToRecipient(
-                  stack.get(Envelope.DataComponents.MAIL_SENDER),
-                  stack.get(Envelope.DataComponents.MAIL_RECIPIENT));
+            @Nullable Address sender = stack.get(Envelope.DataComponents.MAIL_SENDER);
+            @Nullable Address recipient = stack.get(Envelope.DataComponents.MAIL_RECIPIENT);
 
-            if (senderToRecipient != null) {
-                consumer.accept(senderToRecipient);
+            if (Screen.hasShiftDown()) {
+                if (sender != null) {
+                    consumer.accept(Component.translatable("gui.envelope.mail.sender").withStyle(ChatFormatting.GRAY)
+                          .append(": ").withStyle(ChatFormatting.GRAY)
+                          .append(sender.format().asSender().toComponent()));
+                }
+                if (recipient != null) {
+                    consumer.accept(Component.translatable("gui.envelope.mail.recipient").withStyle(ChatFormatting.GRAY)
+                          .append(": ").withStyle(ChatFormatting.GRAY)
+                          .append(recipient.format().asRecipient().toComponent()));
+                }
+            } else {
+                @Nullable Component senderToRecipient = AddressFormatter.senderToRecipient(sender, recipient);
+                if (senderToRecipient != null) {
+                    consumer.accept(senderToRecipient);
+                }
             }
         }
     }